CRUX/requirements.txt
astropy==5.0.1
astroquery==0.4.5
beautifulsoup4==4.10.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
cryptography==36.0.1
cycler==0.11.0
docutils==0.18.1
fitdecode==0.10.0
fitparse==1.2.0
fonttools==4.29.1
html5lib==1.1
idna==3.3
importlib-metadata==4.11.1
jeepney==0.7.1
joblib==1.1.0
keyring==23.5.0
Kivy==2.1.0
Kivy-examples==2.1.0
Kivy-Garden==0.1.4
kivymd==0.104.2
kiwisolver==1.3.2
matplotlib==3.5.1
numpy==1.22.2
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.1
Pillow==9.0.1
pycparser==2.21
pyerfa==2.0.0.1
Pygments==2.11.2
PyOpenGL==3.1.6
pyparsing==3.0.7
PyQt5==5.15.6
PyQt5-Qt5==5.15.2
PyQt5-sip==12.9.1
python-dateutil==2.8.2
pytz==2021.3
pyvo==1.3
PyYAML==6.0
requests==2.27.1
scikit-learn==1.0.2
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
soupsieve==2.3.1
threadpoolctl==3.1.0
urllib3==1.26.9
webencodings==0.5.1
zipp==3.7.0
